+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # CHANGELOG
 
 ## Unreleased
-
 ### Added
 - add evaluation of `Expr` in `Solution`.
+
+## 3.1.0 - 2020-12-17
+### Added
+- add more SCIP functions: `getNSols`, `createPartialSol`
+
+### Fixed
+- consistent handling of filenames for reading/writing
+- fix error when no SCIPOPTDIR env var is set on Windows
 
 ## 3.0.4 - 2020-10-30
 ### Added